requirements.txt
PyQt6>=6.4.0
opencv-python>=4.7.0
ffmpeg-python>=0.2.0
numpy>=1.23.0
pillow>=9.3.0
moviepy>=1.0.3
pytest>=7.3.1
black>=22.3.0
mypy>=1.0.0
